pyproject.toml
[tool.poetry]
name = "availability-checker"
version = "0.1.0"
description = "To achieve the functionality of returning real-time availability of professionals and updating based on current activity or schedule, the following steps and best practices should be implemented using the selected tech stack: Python, FastAPI, PostgreSQL, and Prisma. "
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
bcrypt = "^3.2.0"
fastapi = "^0.75.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
prisma = "*"
pydantic = "*"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
